[arcane,hdf5] Déplace 'HdfVariableInfoBase' de 'arcane/std' vers 'arcane/hdf5'.
