+++ arcane/src/arcane/hdf5/Hdf5VariableInfoBase.cc
@@ -1,11 +1,11 @@
 ﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
 //-----------------------------------------------------------------------------
-// Copyright 2000-2022 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// Copyright 2000-2023 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0
 //-----------------------------------------------------------------------------
 /*---------------------------------------------------------------------------*/
-/* Hdf5ItemVariableInfo.cc                                     (C) 2000-2020 */
+/* Hdf5ItemVariableInfo.cc                                     (C) 2000-2023 */
 /*                                                                           */
 /* Lecture de variables au format HDF5.                                      */
 /*---------------------------------------------------------------------------*/
@@ -18,7 +18,7 @@
 #include "arcane/utils/NotSupportedException.h"
 #include "arcane/utils/ArgumentException.h"
 
-#include "arcane/std/Hdf5VariableInfoBase.h"
+#include "arcane/hdf5/Hdf5VariableInfoBase.h"
 
 #include "arcane/MeshVariable.h"
 #include "arcane/IItemFamily.h"
