+++ arcane/src/arcane/hdf5/srcs.cmake
@@ -1,4 +1,6 @@
 set(ARCANE_SOURCES
   Hdf5Utils.cc
   Hdf5Utils.h
+  Hdf5VariableInfoBase.cc
+  Hdf5VariableInfoBase.h
 )
